Refactors shared HTTP request method into its own class
